+++ lib/index.ts
@@ -2,5 +2,9 @@ import { Task } from './Task';
 import { TaskListHorizontal } from './TaskListHorizontal';
 import { TaskListVertical } from './TaskListVertical';
 import { TaskHorizontal } from './TaskHorizontal';
+import { time_store } from './TimeStore';
 
-export { Task, TaskListHorizontal, TaskListVertical, TaskHorizontal };
+export {
+  Task, TaskListHorizontal,
+  TaskListVertical, TaskHorizontal, time_store
+};